+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Added
 
+- A document
+with [expanded](https://github.com/monadosquito/unpath#expand) file paths can be [saved](https://github.com/monadosquito/unpath#save)
+into a \<document\_path\> file by passing the `--save` (`-s`) flag.
 - Some custom prefixes and suffixes can be set
 by passing them as values of corresponding \<prefix\> or \<suffix\> options.
 - A \<root\_directory\_path\> directory can be omitted from markers file paths.